Fix macos deploy

+++ scripts/1_clean_nodes.sh
@@ -1,4 +1,13 @@
 nodes=$(docker ps -a --format "{{.Names}}" | grep node | sort -n)
-docker stop $nodes
-docker rm $nodes
-docker rmi -f $nodes
+if [[ "$nodes" == "" ]] ;then
+  echo "nodes containers not found"
+else
+  docker stop $nodes
+  docker rm $nodes
+fi
+#docker rmi -f $nodes
+if [[ "$(docker images p2p-bridge_node | grep p2p-bridge_node)" == "" ]];then
+  echo "nodes images not found"
+else
+  docker rmi -f p2p-bridge_node
+fi
